Fix a deadlock problem that troubled me for MONTHS
The main change is in `batch_make.py`. The original design is:
- Spawn a child process to iterate over Makefiles and call make, sending
  results (one MakefileEntry at a time) back to the main process through
  a queue.
- The main process periodically fetches items from the queue, and keeps
  track of the elapsed time. When the time limit is exceeded, the child
  process is terminated and what's left in the queue are retrieved.

The problem is: Queues are implemented based on Pipes, which is a
cross-process connection without buffering. The buffer is kept within
the queue, and a "feed thread" is responsible for: taking stuff out of
the queue buffer, pickling it, and feeding it through the pipe. Of
course, a couple of locks are involved.

Since pipes have no buffer, it's not possible to "get" from a queue
without the child process cooperating. If you try to join a child
process before emptying the queue, you're likely to encounter deadlock
because the child process will not terminate before the feed thread
terminates, which will not happen before the queue is emptied.
Similarly, when "terminate" is called on a child process, it's possible
that the feed thread is holding the lock, which blocks "get" methods
regardless of whether "block==True".

The fix here does not solve the problem entirely: all we've done is try
to empty the queue before calling "terminate", but it's possible that
the child process pushed new stuff into the queue. A better
implementation would be to send a message to the child process asking it
to stop, then try to join it, and only terminate it when it doesn't
respond. However, the current impl. is probably good enough for most
cases.